feat(integrationCenter): mv validate from form wrap to out side
